+++ src/util.js
@@ -1,7 +1,14 @@
 const fs = require('fs-extra');
 const got = require('got');
 const crypto = require('crypto');
+const gmail = require('gmail-send');
+const Stripe = require('stripe');
+const database = require('./database');
 const logger = require('./logger');
+const config = require('../config.json');
+
+const stripe = new Stripe(config.stripe.secret_key);
+const sendGmail = gmail(config.gmail);
 
 function fullUrl(request) {
 	return `${request.protocol}://${request.hostname}${request.originalUrl}`;
@@ -68,11 +75,79 @@ function nintendoPasswordHash(password, pid) {
 	return hashed;
 }
 
+async function handleStripeEvent(event) {
+	if (event.type === 'customer.subscription.updated' || event.type === 'customer.subscription.deleted') {
+		const subscription = event.data.object;
+		const product = await stripe.products.retrieve(subscription.plan.product);
+		const customer = await stripe.customers.retrieve(subscription.customer);
+
+		if (!customer.metadata.pnid_pid && subscription.status !== 'canceled' && subscription.status !== 'unpaid') {
+			// No PNID PID linked to customer. Abort and refund!
+			await stripe.subscriptions.del(subscription.id);
+
+			const invoice = await stripe.invoices.retrieve(subscription.latest_invoice);
+			await stripe.refunds.create({
+				payment_intent: invoice.payment_intent
+			});
+
+			await sendGmail({
+				to: customer.email,
+				subject: 'Pretendo Subscription Failed - No Linked PNID',
+				text: `Your recent subscription to Pretendo Network has failed.\nThis is due to no PNID PID being linked to the Stripe customer account used. The subscription has been canceled and refunded. Please contact Jon immediately.\nStripe Customer ID: ${customer.id}`
+			});
+
+			return;
+		}
+
+		const pid = Number(customer.metadata.pnid_pid);
+		const pnid = await database.PNID.findOne({ pid });
+
+		if (!pnid && subscription.status !== 'canceled' && subscription.status !== 'unpaid') {
+			// PNID does not exist. Abort and refund!
+			await stripe.subscriptions.del(subscription.id);
+
+			const invoice = await stripe.invoices.retrieve(subscription.latest_invoice);
+			await stripe.refunds.create({
+				payment_intent: invoice.payment_intent
+			});
+
+			await sendGmail({
+				to: customer.email,
+				subject: 'Pretendo Subscription Failed - PNID Not Found',
+				text: `Your recent subscription to Pretendo Network has failed.\nThis is due to the provided PNID not being found. The subscription has been canceled and refunded. Please contact Jon immediately.\nStripe Customer ID: ${customer.id}\nPNID PID: ${pid}`
+			});
+
+			return;
+		}
+
+		const updateData = {};
+
+		if (product.metadata.beta === 'true') {
+			switch (subscription.status) {
+				case 'active':
+					updateData.access_level = 1;
+					break;
+
+				case 'canceled': // Subscription was canceled
+				case 'unpaid': // User missed too many payments
+					updateData.access_level = 0;
+					break;
+
+				default:
+					break;
+			}
+
+			await database.PNID.updateOne({ pid }, { $set: updateData }, { upsert: true }).exec();
+		}
+	}
+}
+
 module.exports = {
 	fullUrl,
 	getLocale,
 	apiGetRequest,
 	apiPostGetRequest,
 	apiDeleteGetRequest,
-	nintendoPasswordHash
+	nintendoPasswordHash,
+	handleStripeEvent
 };
